refactor : common 도메인 분리
