+++ blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/common/exception/CustomJwtException.java
@@ -1,4 +1,4 @@
-package com.sparta.blackwhitedeliverydriver.exception;
+package com.sparta.blackwhitedeliverydriver.common.exception;
 
 public class CustomJwtException extends RuntimeException {
     public CustomJwtException(String message) {
